extracted content from changeEmail in separate component
